+++ domain/Episode.py
@@ -11,16 +11,16 @@ class Episode(Base):
     __tablename__ = 'episodes'
 
     id = Column(postgresql.UUID(as_uuid=True), primary_key=True, default=uuid4)
-    bangumi_id = Column(postgresql.UUID(as_uuid=True), ForeignKey('bangumi.id'))
-    bgm_eps_id = Column(Integer)
-    episode_no = Column(Integer)
-    name = Column(TEXT)
+    bangumi_id = Column(postgresql.UUID(as_uuid=True), ForeignKey('bangumi.id'), nullable=False)
+    bgm_eps_id = Column(Integer, nullable=False)
+    episode_no = Column(Integer, nullable=False)
+    name = Column(TEXT, nullable=True)
     name_cn = Column(TEXT, nullable=True)
     duration = Column(String, nullable=True)
     airdate = Column(DATE, nullable=True)
-    status = Column(Integer)
-    create_time = Column(TIMESTAMP, default=datetime.now())
-    update_time = Column(TIMESTAMP, default=datetime.now())
+    status = Column(Integer, nullable=False)
+    create_time = Column(TIMESTAMP, default=datetime.now(), nullable=False)
+    update_time = Column(TIMESTAMP, default=datetime.now(), nullable=False)
 
     bangumi = relationship('Bangumi', back_populates='episodes')