+++ domain/InviteCode.py
@@ -8,4 +8,4 @@ class InviteCode(Base):
     __tablename__ = 'invite_code'
 
     code = Column(postgresql.UUID(as_uuid=True), primary_key=True, default=uuid4)
-    used_by = Column(postgresql.UUID(as_uuid=True))
+    used_by = Column(postgresql.UUID(as_uuid=True), nullable=True)